Resizing the game's world
In this codification section:
I resizes the world of the game (I configured new limits for game's screen) and I configured a camera to follow the player while he moves in the screen.

I also adjusted the player's jump, so that he can only jump if he is on top of something.